feat(wayfinder): update functions and routing logic, add helpers for routing and creating clients
Leaving room for supporting other clients, and consolidating the redirect and resolution logic.
