+++ src/common/wayfinder/routers/index.ts
@@ -13,16 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './error.js';
-export * from './logger.js';
-export * from './ant.js';
-export * from './ant-registry.js';
-export * from './ant-versions.js';
-export * from './faucet.js';
-
-// ao
-export * from './io.js';
-export * from './contracts/ao-process.js';
-
-// wayfinder
-export * from './wayfinder/wayfinder.js';
+export * from './priority.js';
+export * from './random.js';
+export * from './fixed.js';
