+++ src/types/wayfinder.ts
@@ -13,16 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './error.js';
-export * from './logger.js';
-export * from './ant.js';
-export * from './ant-registry.js';
-export * from './ant-versions.js';
-export * from './faucet.js';
+export type WayfinderRouterName = 'random' | 'priority' | 'fixed';
+export interface WayfinderRouter {
+  readonly name: WayfinderRouterName;
+  getTargetGateway: () => Promise<URL>;
+}
 
-// ao
-export * from './io.js';
-export * from './contracts/ao-process.js';
-
-// wayfinder
-export * from './wayfinder/wayfinder.js';
+export type AnyFunction = (...args: any[]) => any;
+export type WayfinderHttpClient<T extends AnyFunction> = T;
